src/requirements.txt
praw==7.7.1
pandas==2.0.3
python-dotenv==1.0.0
textblob==0.17.1
requests==2.31.0